content/media/test/test_bug468190.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=468190
-->
<head>
  <title>Test for bug 468190</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=468190">Mozilla Bug 468190</a>
<pre id="test">
<script class="testbody" type="text/javascript">
// Test if video can be replayed after ended.
var completed = false;
var playingCount = 0;
var endCount = 0;

function startTest() {
  if (completed)
    return false;

  var v = document.getElementById('v');
  v.play();
  return false;
}

function playbackStarted() {
  if (completed)
    return false;

  playingCount++;
}

function playbackEnded() {
  if (completed)
    return false

  endCount++;
  var v = document.getElementById('v');
  ok(v.currentTime >= 3.9 && v.currentTime <= 4.0,
     "Checking currentTime at end: " + v.currentTime);
  ok(v.ended, "Checking playback has ended");
  ok(playingCount > 0, "Expect at least one playing event");
  playingCount = 0;
  if (endCount < 2) {
    v.currentTime = 0;
    ok(v.seeking, "Checking that seeking is now true");
    ok(!v.ended, "Checking ended is no longer set as seeking has begun");
    v.play();
  } else {
    ok(endCount == 2, "Check playback after ended event");
    completed = true;
    SimpleTest.finish();
  }
  return false;
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
<video id='v'
       src='seek.ogv'
       onloadedmetadata='return startTest();'
       onplaying='return playbackStarted();'
       onended='return playbackEnded();'></video>
</body>
</html>
